+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/dependencies/DependencyManager.java
@@ -22,8 +22,10 @@ import com.google.common.io.ByteStreams;
 
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.utils.dependencies.classloader.IsolatedClassLoader;
+import io.github.scroojalix.npcmanager.utils.dependencies.classloader.ReflectionClassLoader;
 import io.github.scroojalix.npcmanager.utils.dependencies.relocation.Relocation;
 import io.github.scroojalix.npcmanager.utils.dependencies.relocation.RelocationHandler;
+import io.github.scroojalix.npcmanager.utils.storage.StorageType;
 
 public class DependencyManager {
 
@@ -33,10 +35,16 @@ public class DependencyManager {
     private final Map<ImmutableSet<Dependency>, IsolatedClassLoader> loaders = new HashMap<>();
     private final EnumMap<Dependency, Path> loaded = new EnumMap<>(Dependency.class);
     private RelocationHandler relocationHandler = null;
+    private ReflectionClassLoader pluginClassLoader;
     
     public DependencyManager(NPCMain main) {
         this.main = main;
         url = "https://repo1.maven.org/maven2/";
+        this.pluginClassLoader = new ReflectionClassLoader(main);
+    }
+
+    public void loadStorageDependencies(StorageType type) {
+        loadDependencies(DependencyRegistry.resolveStorageDependencies(type));
     }
 
     public void loadDependencies(Set<Dependency> dependencies) {
@@ -57,8 +65,8 @@ public class DependencyManager {
 
         this.loaded.put(dependency, file);
 
-        if (this.shouldAutoLoad(dependency)) {
-            this.main.getPluginClassLoader().addJarToClasspath(file);
+        if (DependencyRegistry.shouldAutoLoad(dependency)) {
+            this.pluginClassLoader.addJarToClasspath(file);
         }
     }
 
@@ -78,18 +86,6 @@ public class DependencyManager {
 
         getRelocationHandler().remap(normalFile, remappedFile, rules);
         return remappedFile;
-    }
-
-    private boolean shouldAutoLoad(Dependency dependency) {
-        switch (dependency) {
-            // all used within 'isolated' classloaders, and are therefore not relocated.
-            case ASM:
-            case ASM_COMMONS:
-            case JAR_RELOCATOR:
-                return false;
-            default:
-                return true;
-        }
     }
 
     private synchronized RelocationHandler getRelocationHandler() {