+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/StorageFactory.java
@@ -5,20 +5,28 @@ import java.util.logging.Level;
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.StorageImplementation;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.json.JsonStorage;
+import io.github.scroojalix.npcmanager.utils.storage.implementation.mongodb.MongoStorage;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.mysql.MySQLStorage;
 
 public class StorageFactory {
     
     private NPCMain main;
+    private StorageType type;
 
     public StorageFactory(NPCMain main) {
         this.main = main;
+        this.type = StorageType.parse(main.getConfig().getString("save-method"));
+    }
+
+    public StorageType getType() {
+        return this.type;
     }
 
     public Storage getInstance() {
-        StorageType type = StorageType.parse(main.getConfig().getString("save-method"));
         this.main.log(Level.INFO, "Loading storage provider... [" + type.getName() + "]");
-        return new Storage(this.main, createNewImplementation(type));
+        Storage storage = new Storage(this.main, createNewImplementation(type));
+        storage.init();
+        return storage;
     }
 
     public StorageImplementation createNewImplementation(StorageType type) {
@@ -27,6 +35,8 @@ public class StorageFactory {
                 return new JsonStorage(main);
             case MYSQL:
                 return new MySQLStorage(main);
+            case MONGODB:
+                return new MongoStorage(main);
             default:
                 throw new RuntimeException("Unknown storage type: " + type);
         }