+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/implementation/StorageImplementation.java
@@ -1,8 +1,5 @@
 package io.github.scroojalix.npcmanager.utils.storage.implementation;
 
-import java.util.Set;
-
-import io.github.scroojalix.npcmanager.utils.dependencies.Dependency;
 import io.github.scroojalix.npcmanager.utils.npc.NPCData;
 
 public interface StorageImplementation {
@@ -11,12 +8,9 @@ public interface StorageImplementation {
 
     public void shutdown() throws Throwable;
 
-    public void saveNPC(NPCData data);
+    public void saveNPC(NPCData data) throws Throwable;
 
-    public void removeNPC(String name);
+    public void removeNPC(String name) throws Throwable;
     
-    public void restoreNPCs();
-
-    public Set<Dependency> getDependencies();
-
+    public void restoreNPCs() throws Throwable;
 }